worker: fetch_auth: Send credentials
By default the Fetch API doesn't send any cookies, even for same-origin
requests [1].  This presents a problem when using the fetch API against a
web framework that relies on cookies for authentication (like Django).

Note that the worker runtime is typically not used from the browser, except
when needed to work around issues with conflicts between multiple versions
of Pusher existing on the same page.  One example of when this happens is
if a third-party widget is bundling Pusher, and the main app is also trying
to use it.  For example: [2].

Without this patch, it's not possible to set up authenticated connections
against a Django app with CSRF protection enabled while using the worker
runtime.  With the patch, all's well.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials
[2] https://github.com/pusher/pusher-js/issues/158#issuecomment-274242084