Make key retrieval async
In order to decrypt or encrypt data, the encryption key is required. The
key is only available after the auth endpoint has returned. We need to
wait until the key is available until we send encrypted events.

The async element means that we can't use `trigger` and must create a
new method: `triggerEncrypted`