[WIP] buffering encryoted events post key change
Things to think about:

- At the moment we just drain the buffer, but what happens if the buffer
contains events we can't decrypt, we only want to remove items from the
buffer if we can decrypt them.

- The problem with this is we may end up with events we'll never be able
to decrypt if we "missed" a key

- There needs to be some decoupling of new message and buffer message
handling. Do we want to reauth if we can't decrypt an event from the
buffer? We know we have the right key if we successfully decrypt a new
event?