Clarify some event/message semantics
There were some confusing types concerning messages/events, I've tried
to clarify it a little here. In addition, I've moved everything to do
with Metadata into the Channel (and extended) types.

- Channels *always* emit metadata, but it's usually empty. Although this
seems unpleasant, it hopefully reduces the likelihood of 'Cannot read
property user_id of undefined' errors

- The Connection receives a MessageEvent from the websocket and decodes
it into a PusherEvent which matches with the definition on the protocol
page