Refactor the mocking of the dependency loader
Instead of replacing the exported object with a mock, which works only
because of TS outputting CommonJS code and so webpack not seeing ES6
semantic for the export, the tests are now mocking the different methods
of the object instead, preserving the object identity.
This makes them compatible with the semantic of ES6 named exports.