Add an error log level and switched some warn logging to use it
The potentially contriversial one might be the logging of errors
due to the callback on errors emitted from the connection manager.
While errors can be emitted due to malformed messages (which I
think should be an error), they are also emitted for disconnects
(which I think should be a warning, since the library reconnects).
I decided to log these as warnings so as not to alarm the user
if the SDK disconnects, e.g. due to a server restart.