make some of the configuration easier
Previously we received an options object (<any>) from the caller of the
library and then made an internal config object based on the options
config and some defaults.

We pass some combination of all of these different types of config
around to different things. It seems like we can lean on typescript a
lot more. This commit starts us down that path.

- define an Options interface with all of the configurable parameters
the user can provide. The Options interface should only be used to
override the defaults in the Config interface

- define a Config interface which contains all of the keys of the options
interface. The Config has several required parameters which are
populated from the defauls, *if* they're not provided by the user

- simplify the population of the defaults. There seemed to be some
unnecessary steps

- fix some of the  broken key names (not clear to me right now whether
this will be a breaking change

- use type annotations when we use the Config/Options types

- add a separate Config argument to the ConnectionManager constructor
alongside the ConnectionManagerOptions argument. Previously we just
combined both into the same type and used Collections.extend to force
the type

- add a separate Config argument to the getDefaultStrategy function and
annotate the StrategyOptions type. It's nicer to keep these separate
rather than combining them