Upgrade Karma and Jasmine
Jasmine has some syntax changes after version 2 and its internal clock
behaviour. https://jasmine.github.io/2.0/upgrading.html

I didn't want to touch the code, so you will see some awkward fixes with
this changes.

1) undefined properties are not matched with toEqual
toEqual now doesn't ignore undefined properties and it turns out that we
had a few of those. The obvious(better) fix is to change the code, but
was out of scope for this changes. Therefore, what I did was just adding
them to the expectations.

2) Tests are executed in random order by default
I had to disable this feature because some of our tests rely on their
order to work. (e.g. spec/javascripts/unit/core/logger_spec.js)

3) runs, waitsFor, and waits methods have been removed
I had to backport waitsFor because many tests use this feature to
pause/lock the execution of the test case.
It's a super basic implementation for waitsFor, but adequate enough for
our use.