Update deps to resolve audit errors
It seems that all the recent versions of the `object-assign-deep` break
webpack because they touch objects. That's why I had to replace it with
the `webpack-merge` lib.

There are a few more outdated dependencies, but I didn't want to include
them with these changes.

I didn't bump webpack to latest major version becaues it requires
configuration changes.