Increase Jasmine default timeout in a few places
There are a few integration tests that timeout sometimes after Jasmine
upgrade. Jasmine 2+ introduced global timeout of 5secs for test cases.

This one is tricky to reproduce. I saw it once in Github action, and I
could reproduce it after running the test many times on local machine.

I didn't look into the actual implementation since it seems to be a
minor issue. That's why I just increased the default timeout in a couple
of places.