Use defaults for auth when objects are not fully provided (#8)
We have defaults for auth endpoints and transport. However, they were not being picked up if the caller provided one of the parameters but not the other in the object. That meant that users could either provide no parameters or all of them. This PR allows them to pass any subset of the parameters relying on defaults for the other ones.